Inline some classes that just added confusion given that they were only needed by conversion functions that did most of the work (within the scope of XMF/ANI, anyway)
Fix many inspection results.

Add numeric_cast to prevent headaches later.

+++ X4ConverterTools/include/X4ConverterTools/xmf/XmfDataBuffer.h
@@ -24,9 +24,9 @@ namespace xmf {
 
         bool IsVertexBuffer() const;
 
-        int GetCompressedDataSize() const;
+        uint64_t GetCompressedDataSize() const;
 
-        int GetUncompressedDataSize() const;
+        uint64_t GetUncompressedDataSize() const;
 
         uint8_t *GetData();
